Refactored event, which must be shared between bots and observers.

+++ robocode2-server/src/main/java/net/robocode2/model/events/BotDeathEvent.java
@@ -8,7 +8,7 @@ public final class BotDeathEvent implements Event {
 		this.botId = botId;
 	}
 
-	public double getBotId() {
+	public int getBotId() {
 		return botId;
 	}
 }
